+++ magik-typed-lint/src/main/java/nl/ramsolutions/sw/magik/typedlint/package-info.java
@@ -1,0 +1,4 @@
+/** Defaults. */
+@edu.umd.cs.findbugs.annotations.DefaultAnnotationForParameters(
+    edu.umd.cs.findbugs.annotations.NonNull.class)
+package nl.ramsolutions.sw.magik.typedlint;